Many changes.
In an attempt to keep this library simple, I have removed some functionality.
The reason for this is that it should be used as an intermediary library.
Other solutions with more functionality can be built on top of it.
My goal is for this to be a tool with the single purpose of extracting prices, thats all.
Prices can and should be stored and read from a db.

Fix: Price::as_u32() was changed to Price::as_i32().
     This was a crucial fix becase prices can actually be negative (I totally forgot..).

Change: Instead of returning HourlyResult<Price>, we only return all prices in a Vec<Price>.
        In short, either an empty Vec<Price> is returned, or all prices for the day

Change: Datetime in elspot::Price is now DateTime<Utc> instead of NaiveDateTime.
        This is better for preserving correct time data and utc is the easiest option.
        The struct fields 'from' and 'to' are also now private, I have provided struct methods.
        Keep in mind that if you want to save prices to for example a db,
        then saving times as both the utc timestamp plus a naivedate string "YYYY-MM-DD" is good..
        The utc datetime can have a different date due to tz offset,
        this means that keeping just the utc date itself might make db queries harder to write.
        Use elspot::hourly::Hourly::date(). It will return the date (same for all prices).
        It is easy to correct the offset from a utc datetime once read from db when you know the region.
        For example, by using function: region_time::dt_region_from_utc_dt(&utc_td, &region).

Add: Some extra validation when extracting prices because we now preserve correct time data.
     New datatype region_times::PriceHours will give us the expected number of prices.

Add: Check if prices are official, elspot::hourly::Hourly::prices_are_official().
     Good for checking whether or not the prices might change.
     For now (maybe forever..), it will only return 'true' if every single price in the dataset is marked official.

Remove: No longer possible to extract one price from a region.
        Removed elspot::hourly::Hourly::price_for_region_at_utc_dt().
        Removed elspot::hourly::Hourly::price_now_for_region().
        We extract all prices or nothing.
        The idea behind it is that we should really only do it once, ever..; see comment on top.

Tests: They are changed/updated to keep up with the changing codebase.
       I removed tests/common/mod.rs because I think there are to few tests.
       At this time, there is no need for test code being split up in modules.

Note: Some commented out code remains.
      I might use it later, so I will just keep it..

Todo: Write more tests.
      Check that datetimes for prices matches for both region tz and utc.